[versit] Handle malformed concatentations of vCard files
If vCard documents which do not end in \r\n are concatenated together,
the result is a malformed vCard document with a line like
END:VCARDBEGIN:VCARD\r\n
in it.  This commit adds logic to the parser to detect those types of
lines and replace them with
END:VCARD\r\nBEGIN:VCARD\r\n
lines.

Signed-off-by: Chris Adams <chris.adams@jollamobile.com>